# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Access-Nex
  description: >
    Self-hosted OAuth2 / OpenID Connect provider. This document covers the
    JSON surface a client application (including a separately-hosted
    frontend — see docs/FRONTEND.md) can build against: the standard
    OIDC/OAuth2 endpoints, the /api/v1 account API, and the /api/admin
    management API. The server-rendered HTML pages (/login, /portal, /admin,
    /device) are not part of this contract — they exist for zero-setup
    browser use and can be replaced entirely by a custom frontend calling
    these APIs.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Default local install
security:
  - sessionCookie: []
paths:
...
  /api/v1/login:
    post:
      summary: Password login (step 1)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { username: { type: string }, password: { type: string } }
      responses:
        "200":
          description: Signed in, or a 2FA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties: { status: { type: string, enum: [signed_in] } }
                  - type: object
                    properties:
                      status: { type: string, enum: [2fa_required] }
                      token: { type: string, description: "pass to /api/v1/login/totp or /login/webauthn/*" }
                      methods: { type: array, items: { type: string, enum: [totp, webauthn] } }
        "401": { description: Invalid credentials }
        "429": { description: Rate limited }
  /api/v1/login/totp:
    post:
      summary: Password login (step 2, TOTP)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { token: { type: string }, code: { type: string } }
      responses:
        "200": { description: Signed in; sets the session cookie }
        "401": { description: Invalid code }
...
  /api/v1/grants/{client_id}:
    delete:
      summary: Revoke consent for an application
      parameters: [{ name: client_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Revoked } }
...
  /api/v1/sessions/{id}:
    delete:
      summary: End a session (sign out that device)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Revoked } }
...
  /api/v1/identities/{id}:
    delete:
      summary: Unlink an external identity
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Unlinked }
        "409": { description: Refused — this is the account's only sign-in method }
  /api/v1/password:
    post:
      summary: Change password (while signed in)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { current_password: { type: string }, new_password: { type: string, minLength: 8 } }
      responses:
        "200": { description: Changed }
        "401": { description: Current password incorrect }
...
  /api/v1/forgot-password:
    post:
      summary: Request a password reset email
      security: []
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { email: { type: string } } }
      responses:
        "200":
          description: >
            Always returns { status: if_exists_email_sent } regardless of
            whether the address matched an account, to avoid user enumeration.
  /api/v1/reset-password:
    post:
      summary: Complete a password reset (token from the emailed link)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { token: { type: string }, new_password: { type: string, minLength: 8 } }
      responses:
        "200": { description: Changed }
        "400": { description: Link invalid or expired }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oidc_sid
      description: >
        Set by /login, /api/v1/login(+totp), or /login/webauthn/finish.
        Mutating /api/v1 and /api/admin requests additionally require the
        X-Access-Nex-Api / X-Access-Nex-Admin header respectively.
    clientBasic:
      type: http
      scheme: basic
      description: OAuth2 client_id:client_secret (confidential clients); public clients omit auth entirely
    bearerToken:
      type: http
      scheme: bearer
    dpopToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'For DPoP-bound tokens: "Authorization: DPoP <token>" plus a DPoP proof header'